zonemd, zonemd unit test in own file.
